Move out common apis and irq configuration from ioapic.c
Move out the ioapic apis which are for irq ops
