+++ hypervisor/Makefile
@@ -200,6 +200,7 @@ BOOT_C_SRCS += arch/x86/rtcm.c
 # hardware management component
 HW_S_SRCS += arch/x86/idt.S
 HW_C_SRCS += arch/x86/ioapic.c
+HW_C_SRCS += arch/x86/ioapic_irq.c
 HW_C_SRCS += arch/x86/lapic.c
 HW_C_SRCS += arch/x86/cpu.c
 HW_C_SRCS += arch/x86/cpu_caps.c
