+++ hypervisor/include/arch/x86/ioapic.h
@@ -8,6 +8,14 @@
 #define IOAPIC_H
 
 #include <x86/apicreg.h>
+#include <types.h>
+#include <x86/lib/spinlock.h>
+#include <x86/ioapic.h>
+#include <x86/io.h>
+#include <logmsg.h>
+
+extern uint8_t ioapic_num;
+extern spinlock_t ioapic_lock;
 
 #define NR_LEGACY_IRQ		16U
 #define NR_LEGACY_PIN		NR_LEGACY_IRQ
@@ -19,6 +27,8 @@ struct ioapic_info {
 	uint32_t nr_pins;	/* Number of Interrupt inputs as determined by Max. Redir Entry Register */
 };
 
+extern struct ioapic_info ioapic_array[CONFIG_MAX_IOAPIC_NUM];
+
 void ioapic_setup_irqs(void);
 void init_ioapic();
 
@@ -28,6 +38,49 @@ int32_t init_ioapic_id_info(void);
 uint8_t ioapic_irq_to_ioapic_id(uint32_t irq);
 
 uint8_t get_platform_ioapic_info (struct ioapic_info **plat_ioapic_info);
+void *map_ioapic(uint64_t ioapic_paddr);
+void ioapic_get_rte_entry(void *ioapic_base, uint32_t pin, union ioapic_rte *rte);
+
+static inline uint32_t
+ioapic_read_reg32(void *ioapic_base, const uint32_t offset)
+{
+	uint32_t v;
+	uint64_t rflags;
+
+	spinlock_irqsave_obtain(&ioapic_lock, &rflags);
+
+	/* Write IOREGSEL */
+	mmio_write32(offset, ioapic_base + IOAPIC_REGSEL);
+	/* Read  IOWIN */
+	v = mmio_read32(ioapic_base + IOAPIC_WINDOW);
+
+	spinlock_irqrestore_release(&ioapic_lock, rflags);
+	return v;
+}
+
+static inline void
+ioapic_write_reg32(void *ioapic_base, const uint32_t offset, const uint32_t value)
+{
+	uint64_t rflags;
+
+	spinlock_irqsave_obtain(&ioapic_lock, &rflags);
+
+	/* Write IOREGSEL */
+	mmio_write32(offset, ioapic_base + IOAPIC_REGSEL);
+	/* Write IOWIN */
+	mmio_write32(value, ioapic_base + IOAPIC_WINDOW);
+
+	spinlock_irqrestore_release(&ioapic_lock, rflags);
+}
+
+static inline void
+ioapic_set_rte_entry(void *ioapic_base,
+		uint32_t pin, union ioapic_rte rte)
+{
+	uint32_t rte_addr = (pin * 2U) + 0x10U;
+	ioapic_write_reg32(ioapic_base, rte_addr, rte.u.lo_32);
+	ioapic_write_reg32(ioapic_base, rte_addr + 1U, rte.u.hi_32);
+}
 
 /**
  * @defgroup ioapic_ext_apis IOAPIC External Interfaces
